requirements_audio_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
audiofile=0.2.1=pypi_0
audioread=2.1.6=py37_0
audtorch=0.2.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h516909a_0
ca-certificates=2019.5.15=0
certifi=2019.6.16=py37_1
cffi=1.12.3=py37h2e261b9_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py37_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py_0
expat=2.2.6=he6710b0_0
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gettext=0.19.8.1=hc5be6a0_1002
glib=2.56.2=hd408876_0
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py37h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
intel-openmp=2019.4=243
joblib=0.13.2=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=py37he6710b0_0
lame=3.100=h14c3975_1001
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libflac=1.3.1=hf484d3e_1002
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
libogg=1.3.2=h14c3975_1001
libpng=1.6.37=hbc83047_0
librosa=0.7.0=py_0
libsndfile=1.0.28=hf484d3e_1000
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libvorbis=1.3.5=h14c3975_1001
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.29.0=pypi_0
lmdb=0.96=pypi_0
matplotlib=3.1.0=py37h5429711_0
mkl=2019.4=243
mkl-service=2.1.0=py37h516909a_0
mkl_fft=1.0.12=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
msgpack-numpy=0.4.4.2=py37_0
msgpack-python=0.6.1=py37hfd86e86_1
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
ninja=1.9.0=py37hfd86e86_0
numba=0.45.1=py37hb3f55d8_0
numpy=1.16.4=py37h7e9f1db_0
numpy-base=1.16.4=py37hde5b4d6_0
olefile=0.46=py37_0
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1c=h7b6447c_1
pandas=0.25.0=py37he6710b0_0
pcre=8.43=he6710b0_0
pillow=6.1.0=py37h34e0f95_0
pip=19.1.1=py37_0
psutil=5.6.3=py37h7b6447c_0
pycparser=2.19=py37_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py37h05f1152_2
pysoundfile=0.10.2=py_1001
python=3.7.3=h0371630_0
python-dateutil=2.8.0=py37_0
pytorch=1.2.0=py3.7_cuda10.0.130_cudnn7.6.2_0
pytz=2019.1=py_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
resampy=0.2.1=pypi_0
scikit-learn=0.21.2=py37hd81dba3_0
scipy=1.3.0=pypi_0
setuptools=41.0.1=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
soundfile=0.10.2=pypi_0
sox=1.3.7=pypi_0
sqlite=3.29.0=h7b6447c_0
tabulate=0.8.3=pypi_0
tk=8.6.8=hbc83047_0
torchaudio=0.2=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.4.0=py37_cu100
tornado=6.0.3=py37h7b6447c_0
tqdm=4.32.1=py_0
typing=3.6.4=py37_0
wheel=0.33.4=py37_0
x264=1!152.20180806=h14c3975_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
